Replace uint with uint256 in unstakeETH function parameter

+++ StakingContract/Rebase.sol
@@ -1998,7 +1998,7 @@ contract Rebase is ReentrancyGuard {
         require(ERC20(token).transfer(msg.sender, quantity), "Unable to transfer token");
     }
 
-    function unstakeETH(uint quantity, address app) external nonReentrant {
+    function unstakeETH(uint256 quantity, address app) external nonReentrant {
         _unstake(app, _WETH, quantity);
         _getReToken(_WETH).burn(msg.sender, quantity);
         WETH(_WETH).withdraw(quantity);
